chore: improve dx
