+++ packages/dashboard/src/components/usersTable.tsx
@@ -23,7 +23,7 @@ import { immer } from "zustand/middleware/immer";
 import apiRequestHandlerFactory from "../lib/apiRequestHandlerFactory";
 import { useAppStore } from "../lib/appStore";
 import { monospaceCell } from "../lib/datagridCells";
-import { propertiesStore } from "./filterDisplay";
+import { propertiesStore } from "./usersFilter";
 
 export const UsersTableParams = Type.Pick(GetUsersRequest, [
   "cursor",
@@ -150,7 +150,6 @@ export default function UsersTable({
   segmentId,
   direction,
   cursor,
-  userPropertyIds,
   onPaginationChange,
 }: UsersTableProps) {
   const router = useRouter();
@@ -165,7 +164,9 @@ export default function UsersTable({
   const setUsers = usersStore((store) => store.setUsers);
   const setUsersPage = usersStore((store) => store.setUsersPage);
   const setPreviousCursor = usersStore((store) => store.setPreviousCursor);
-  const selectedPropertySelectedValue = propertiesStore((store) => store.selectedPropertySelectedValue);
+
+  // used to filter by property 
+  const selectedPropertySelectedValue = propertiesStore((store) => store.filter);
   const userPropertyFilter = useMemo(() => Object.values(selectedPropertySelectedValue), [selectedPropertySelectedValue])
 
   const usersPage = useMemo(
